Move build message to messaging
