+++ app/routes/funding-options/validation.js
@@ -1,7 +1,7 @@
 const getPollingResponse = require('../../polling')
 const { sendAgreementCalculateMessage } = require('../../messaging')
 const cache = require('../../cache')
-const buildMessage = require('../../calculation')
+const buildMessage = require('../../messaging/create-calculation-message')
 
 module.exports = [{
   method: 'GET',
